// src/generated/server/alsaglobal/conflict/v1/service_server.ts
export interface ListAcledEventsRequest {
  start: number;
  end: number;
  pageSize: number;
  cursor: string;
  country: string;
}

export interface ListAcledEventsResponse {
  events: AcledConflictEvent[];
  pagination?: PaginationResponse;
}

export interface AcledConflictEvent {
  id: string;
  eventType: string;
  country: string;
  location?: GeoCoordinates;
  occurredAt: number;
  fatalities: number;
  actors: string[];
  source: string;
  admin1: string;
}

export interface GeoCoordinates {
  latitude: number;
  longitude: number;
}

export interface PaginationResponse {
  nextCursor: string;
  totalCount: number;
}
// ...
export interface ServerContext {
  request: Request;
  pathParams: Record<string, string>;
  headers: Record<string, string>;
}